Add support for timeout handling
For most of the X change connections we need to maintain timeouts. So lets
make some reusable components here.

The timeout policy object is created by the datagram_receiver and expects
to get a reference to the socket it is keeping track of the timeout for.
If you want to keep track of the timeout for a handler you can create a
type per handler instead.

Signed-off-by: Jan Blunck <jblunck@infradead.org>